requirements.txt
# ─── Core Framework ───
flask>=2.3.0
flask-cors>=4.0.0
flask-jwt-extended>=4.5.0
werkzeug>=2.3.0

# ─── Database ───
pymongo>=4.5.0
dnspython>=2.4.0

# ─── Security ───
bcrypt>=4.0.0
python-dotenv>=1.0.0

# ─── Utilities ───
python-dateutil>=2.8.0
email-validator>=2.0.0
